Renaming C++ files & methods according to recommended naming conventions and aligning them with Python's API.
